frontend/requirements.txt
streamlit==1.45.0
fpdf2>=2.7.0
requests==2.32.3
python-dotenv==1.1.0